TemplateMap: Transform georeferenced data on loading
The TemplateMap functions rely on data being present in the
coordinate system of the owning map if the georeferenced property
is set to true. In TemplateMap, the required transformation of data
was missing, and so support for georeferenced state was incomplete
until now.
In contrast, OgrTemplate already used georeferenced state, did the
transformation on loading, and relied on TemplateMaps functions
supporting georeferenced state.